chore: Refactor with Functions, part III
