+++ src/managers.js
@@ -1,16 +1,20 @@
 import * as GitConfigManager from './managers/GitConfigManager'
 import * as GitIgnoreManager from './managers/GitIgnoreManager'
 import * as GitIndexManager from './managers/GitIndexManager'
+import * as GitObject from './managers/GitObject'
 import * as GitObjectManager from './managers/GitObjectManager'
+import * as GitPktLine from './managers/GitPktLine'
 import * as GitRefManager from './managers/GitRefManager'
 import * as GitRemoteConnection from './managers/GitRemoteConnection'
 import * as GitRemoteHTTP from './managers/GitRemoteHTTP'
 import * as GitRemoteManager from './managers/GitRemoteManager'
 import * as GitShallowManager from './managers/GitShallowManager'
 
+export { GitPktLine }
 export { GitConfigManager }
 export { GitIgnoreManager }
 export { GitIndexManager }
+export { GitObject }
 export { GitObjectManager }
 export { GitRefManager }
 export { GitRemoteConnection }
