+++ src/models.js
@@ -2,9 +2,7 @@ export * from './models/FileSystem'
 export * from './models/GitCommit'
 export * from './models/GitConfig'
 export * from './models/GitIndex'
-export * from './models/GitObject'
 export * from './models/GitPackIndex'
-export * from './models/GitPktLine'
 export * from './models/GitRefSpec.js'
 export * from './models/GitRefSpecSet.js'
 export * from './models/GitTree'
